fix 2

+++ 0x15-file_io/1-create_file.c
@@ -9,7 +9,7 @@
 int create_file(const char *filename, char *text_content)
 {
 	int fd, len = 0, i = 0;
-	size_t num_write;
+	int num_write;
 
 	if (filename == NULL)
 		return (-1);
